requirements.txt
opencv-python
Pillow
pandas
requests
